ml_service/requirements.txt
fastapi
uvicorn[standard]
python-dateutil
mysql-connector-python